--- src/main/java/com/hanghae/prevstudy/domain/member/MemberService.java
@@ -1,7 +1,0 @@
-package com.hanghae.prevstudy.domain.member;
-
-public interface MemberService {
-    void signup(MemberAddRequest memberAddRequest);
-
-    AuthResultDto login(LoginRequest loginRequest);
-}
