+++ src/main/java/com/hanghae/prevstudy/domain/member/entity/Member.java
@@ -1,4 +1,4 @@
-package com.hanghae.prevstudy.domain.member;
+package com.hanghae.prevstudy.domain.member.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
