+++ src/main/java/com/hanghae/prevstudy/domain/member/repository/MemberRepository.java
@@ -1,5 +1,6 @@
-package com.hanghae.prevstudy.domain.member;
+package com.hanghae.prevstudy.domain.member.repository;
 
+import com.hanghae.prevstudy.domain.member.entity.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
