+++ src/main/java/com/hanghae/prevstudy/domain/member/service/MemberService.java
@@ -1,0 +1,11 @@
+package com.hanghae.prevstudy.domain.member.service;
+
+import com.hanghae.prevstudy.domain.member.dto.AuthResultDto;
+import com.hanghae.prevstudy.domain.member.dto.LoginRequest;
+import com.hanghae.prevstudy.domain.member.dto.MemberAddRequest;
+
+public interface MemberService {
+    void signup(MemberAddRequest memberAddRequest);
+
+    AuthResultDto login(LoginRequest loginRequest);
+}
