+++ src/test/java/com/hanghae/prevstudy/domain/member/MemberControllerTest.java
@@ -4,6 +4,13 @@ import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.DeserializationFeature;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.hanghae.prevstudy.domain.member.controller.MemberController;
+import com.hanghae.prevstudy.domain.member.dto.AuthResultDto;
+import com.hanghae.prevstudy.domain.member.dto.LoginRequest;
+import com.hanghae.prevstudy.domain.member.dto.LoginResponse;
+import com.hanghae.prevstudy.domain.member.dto.MemberAddRequest;
+import com.hanghae.prevstudy.domain.member.exception.MemberErrorCode;
+import com.hanghae.prevstudy.domain.member.service.MemberService;
 import com.hanghae.prevstudy.domain.security.TokenDto;
 import com.hanghae.prevstudy.global.exception.GlobalExceptionHandler;
 import com.hanghae.prevstudy.global.exception.PrevStudyException;
